requirements.txt
# Core CV Libraries
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
insightface>=0.7.3
onnxruntime>=1.15.0
# For FastReID - install separately from GitHub if needed
# git+https://github.com/JDAI-CV/fast-reid.git

# Graph & Vector Storage
neo4j>=5.12.0
faiss-cpu>=1.7.4
networkx>=3.1

# API & Services
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
celery>=5.3.0
redis>=5.0.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Analytics & Visualization
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.17.0
streamlit>=1.28.0

# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pillow>=10.0.0
pyyaml>=6.0.1

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Jupyter for experiments
jupyter>=1.0.0
ipykernel>=6.26.0
